Consolidate token form row models into one file
